# Add Support for Preallocating Entry Slice in Batch Verifier