Before:
```
goos: darwin
goarch: arm64
pkg: github.com/hdevalence/ed25519consensus
BenchmarkCreateBatch/1-12  	 5439469	       214.8 ns/op	      96 B/op	       2 allocs/op
BenchmarkCreateBatch/8-12  	  704911	      1714 ns/op	    1104 B/op	      12 allocs/op
BenchmarkCreateBatch/64-12 	   90453	     13074 ns/op	    9680 B/op	      71 allocs/op
BenchmarkCreateBatch/1024-12         	    5697	    204843 ns/op	  155089 B/op	    1035 allocs/op
BenchmarkCreateBatch/4096-12         	    1401	    859972 ns/op	  892369 B/op	    4111 allocs/op
BenchmarkCreateBatch/16384-12        	     316	   3727873 ns/op	 5094872 B/op	   16405 allocs/op
PASS
ok  	github.com/hdevalence/ed25519consensus	10.901s
```

After (**~71% reduction in bytes allocated, ~13% reduction in runtime on batch of 16384**):
```
goos: darwin
goarch: arm64
pkg: github.com/hdevalence/ed25519consensus
BenchmarkCreatePreallocatedBatch/1-12  	 5431094	       213.9 ns/op	      96 B/op	       2 allocs/op
BenchmarkCreatePreallocatedBatch/8-12  	  762730	      1585 ns/op	     704 B/op	       9 allocs/op
BenchmarkCreatePreallocatedBatch/64-12 	   95875	     12557 ns/op	    6144 B/op	      65 allocs/op
BenchmarkCreatePreallocatedBatch/1024-12         	    5967	    199347 ns/op	   90112 B/op	    1025 allocs/op
BenchmarkCreatePreallocatedBatch/4096-12         	    1501	    805202 ns/op	  360449 B/op	    4097 allocs/op
BenchmarkCreatePreallocatedBatch/16384-12        	     374	   3237073 ns/op	 1441795 B/op	   16385 allocs/op
PASS
ok  	github.com/hdevalence/ed25519consensus	10.878s
```